Only delete directory if it exists.
git-svn-id: https://svn.salilab.org/imp/trunk@17007 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ tools/coverage/setup.py
@@ -10,7 +10,8 @@ def clean_gcdas():
                 os.unlink(os.path.join(dirpath, f))
 
 def setup_python_coverage():
-    shutil.rmtree('coverage')
+    if os.path.exists('coverage'):
+        shutil.rmtree('coverage')
     os.mkdir('coverage')
     open('coverage/sitecustomize.py', 'w').write("""
 import coverage
